+++ tiledb/cloud/array.py
@@ -8,9 +8,10 @@ import numpy
 
 from tiledb.cloud import client
 from tiledb.cloud import config
-from tiledb.cloud import results
+from tiledb.cloud import results as results_alias
 from tiledb.cloud import tiledb_cloud_error
 from tiledb.cloud import utils
+from tiledb.cloud._results import results
 from tiledb.cloud.rest_api import ApiException as GenApiException
 from tiledb.cloud.rest_api import models
 
@@ -502,7 +503,7 @@ def apply_base(
     return client.send_udf_call(
         api_instance.submit_udf,
         submit_kwargs,
-        results.Decoder(result_format),
+        results_alias.Decoder(result_format),
         id_callback=_maybe_set_last_udf_id,
         results_stored=store_results,
     )
@@ -648,7 +649,7 @@ def exec_multi_array_udf_base(
     return client.send_udf_call(
         api_instance.submit_multi_array_udf,
         submit_kwargs,
-        results.Decoder(result_format),
+        results_alias.Decoder(result_format),
         id_callback=_maybe_set_last_udf_id,
         results_stored=store_results,
     )
